backend/requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
influxdb-client==1.38.0

# Message Queue
kafka-python==2.0.2
aiokafka==0.8.11

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Data Processing & ML
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
scipy==1.11.4

# Drift Detection
alibi-detect==0.15.0
evidently==0.3.0

# Monitoring & Metrics
prometheus-client==0.19.0
structlog==23.2.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Billing
stripe==7.8.0

# Cloud Providers
boto3==1.34.0
google-cloud-storage==2.10.0
azure-storage-blob==12.19.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3 